Refactor game logic and UI components
- Updated logic.py to include new probability calculations for woodcutting and mining based on player level.
- Enhanced level-up and achievement checking functions to accept player data.
- Introduced new pure functions in logic_pure.py for woodcutting and mining actions, including gem drop mechanics.
- Added unit tests for new functionalities in test_logic.py, covering woodcutting and mining actions.
- Expanded ui.py with new dialogs for tree and ore selection, integrating logic checks for player levels.
- Improved error handling and user feedback in UI dialogs.
- Updated __pycache__ with new compiled Python files reflecting the changes.
