+++ logic.py
@@ -1,5 +1,12 @@
 from aqt import mw
-from .constants import ORE_DATA, EXP_TABLE, ACHIEVEMENTS
+from .constants import (
+    ORE_DATA,
+    EXP_TABLE,
+    ACHIEVEMENTS,
+    BASE_WOODCUTTING_PROBABILITY,
+    BASE_MINING_PROBABILITY,
+    LEVEL_BONUS_FACTOR,
+)
 from .ui import show_level_up_dialog, show_achievement_dialog, update_menu_visibility
 
 # Game logic functions
@@ -51,9 +58,9 @@ def get_exp_to_next_level(player_data, EXP_TABLE):
         return 0
     return EXP_TABLE[current_level] - player_data["total_exp"]
 
-from .logic_pure import calculate_new_level
+from .logic_pure import calculate_new_level, calculate_probability_with_level
 
-def level_up_check(skill):
+def level_up_check(skill, player_data):
     skill_map = {
         "Mining": ("mining_level", "mining_exp"),
         "Woodcutting": ("woodcutting_level", "woodcutting_exp"),
@@ -71,8 +78,28 @@ def level_up_check(skill):
 
 from .logic_pure import get_newly_completed_achievements
 
-def check_achievements():
+def check_achievements(player_data):
     newly_completed = get_newly_completed_achievements(player_data, ACHIEVEMENTS)
     for achievement in newly_completed:
         player_data["completed_achievements"].append(achievement)
         show_achievement_dialog(achievement, ACHIEVEMENTS[achievement])
+
+
+def calculate_woodcutting_probability(player_level: int, tree_probability: float) -> float:
+    return calculate_probability_with_level(
+        player_level,
+        BASE_WOODCUTTING_PROBABILITY,
+        LEVEL_BONUS_FACTOR,
+        tree_probability,
+        cap=0.95,
+    )
+
+
+def calculate_mining_probability(player_level: int, ore_probability: float) -> float:
+    return calculate_probability_with_level(
+        player_level,
+        BASE_MINING_PROBABILITY,
+        LEVEL_BONUS_FACTOR,
+        ore_probability,
+        cap=0.95,
+    )
